research-service-py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
html5lib==1.1
apscheduler==3.10.4
readability-lxml==0.8.1
lxml==5.3.0
